fix: remove callsite from logs
